refactor(checkpoint): unify aborted sink cleanup
